+++ internal/user/model/user.go
@@ -1,6 +1,6 @@
 package model
 
 type User struct {
-	ID   uint32 `json:"id" db:"id"`
+	ID   uint64 `json:"id" db:"id"`
 	Name string `json:"name" db:"name"`
 }
